+++ linstor_client/commands/rsc_cmds.py
@@ -506,6 +506,8 @@ class ResourceCommands(Commands):
                                                             linstor.consts.NAMESPC_AUXILIARY + '/'),
                 replicas_on_different=self.prepare_argparse_list(
                     args.replicas_on_different, linstor.consts.NAMESPC_AUXILIARY + '/'),
+                x_replicas_on_different=self.prepare_argparse_dict_str_int(
+                    args.x_replicas_on_different, linstor.consts.NAMESPC_AUXILIARY + '/'),
                 diskless_on_remaining=self.parse_diskless_on_remaining(args),
                 async_msg=async_flag,
                 layer_list=args.layer_list,