+++ linstor_client/commands/rsc_dfn_cmds.py
@@ -260,6 +260,8 @@ class ResourceDefinitionCommands(Commands):
             replicas_on_same=self.prepare_argparse_list(args.replicas_on_same, linstor.consts.NAMESPC_AUXILIARY + '/'),
             replicas_on_different=self.prepare_argparse_list(
                 args.replicas_on_different, linstor.consts.NAMESPC_AUXILIARY + '/'),
+            x_replicas_on_different=self.prepare_argparse_dict_str_int(
+                args.x_replicas_on_different, linstor.consts.NAMESPC_AUXILIARY + '/'),
             diskless_on_remaining=self.parse_diskless_on_remaining(args),
             layer_list=args.layer_list,
             provider_list=args.providers,