+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Autoplacer: Add --x-replicas-on-different option
+
 ## [1.22.1] - 2024-04-25
 
 ### Changed